requirements.txt
huggingface-hub==0.24.5
sentence-transformers==3.0.1
langchain==0.2.14
langchain-community==0.2.12
langchain-huggingface==0.0.3
langchain-openai==0.1.22
langchain-chroma
python-dotenv==1.0.1
transformers[agent]
